set non-blocking socket flag
Sockets by default are blocking if the send buffer becomes full (frames
not being terminated). One can set the O_NONBLOCK flag using `fcntl()`
to disable this - the application will recieve EAGAIN rather that block
[1](https://rtime.felk.cvut.cz/can/socketcan-qdisc-final.pdf).

Since most Node applications with socketcan are doing things other than
CAN, it makes sense to set this flag. This means if the CAN socket is
not emptying (the device is disconnected for example), the whole
application does not freeze.

[fcntl](http://man7.org/linux/man-pages/man2/fcntl.2.html)